Remove mock utilities and add useGeminiData hook
